Omada widget : spliting widget between WLAN and LAN/WAN fields to have no more than 5 fields per widget.
