+++ src/widgets/components.js
@@ -25,6 +25,7 @@ const components = {
   npm: dynamic(() => import("./npm/component")),
   nzbget: dynamic(() => import("./nzbget/component")),
   omada: dynamic(() => import("./omada/component")),
+  omadalan: dynamic(() => import("./omadalan/component")),
   ombi: dynamic(() => import("./ombi/component")),
   overseerr: dynamic(() => import("./overseerr/component")),
   paperlessngx: dynamic(() => import("./paperlessngx/component")),
