+++ src/widgets/omadalan/component.jsx
@@ -20,20 +20,22 @@ export default function Component({ service }) {
   if (!omadaData) {
     return (
       <Container service={service}>
-        <Block label="omada.connectedAp" />
-        <Block label="omada.activeUser" />
         <Block label="omada.alerts" />
-        <Block label="omada.isolatedAp" />
+        <Block label="omada.connectedGateway" />
+        <Block label="omada.powerConsumption" />
+        <Block label="omada.availablePorts" />
+        <Block label="omada.connectedSwitches" />
       </Container>
     );
   }
 
   return (
     <Container service={service}>
-      <Block label="omada.connectedAp" value={t( "common.number", { value: omadaData.connectedAp})} />
-      <Block label="omada.activeUser" value={t( "common.number", { value: omadaData.activeUser })} />
       <Block label="omada.alerts" value={t( "common.number", { value: omadaData.alerts })} />
-      <Block label="omada.isolatedAp" value={t("common.number", { value: omadaData.isolatedAp})} />
+      <Block label="omada.connectedGateway" value={t("common.number", { value: omadaData.connectedGateways})}/>
+      <Block label="omada.powerConsumption" value={t("common.power", { value: omadaData.powerConsumption})}/>
+      <Block label="omada.availablePorts" value={t("common.number", { value: omadaData.availablePorts})}/>
+      <Block label="omada.connectedSwitches" value={t("common.number", { value: omadaData.connectedSwitches})} />
     </Container>
   );
 }
